requirements.txt
mysql_connector_repackaged==0.3.1
numpy==1.22.3
pandas==1.4.2
progressbar33==2.4
